[WIP] small ref - move error handling to a separate class
